feat: add git status hint to status row
